server/requirements.txt
django==3.2.22
mysqlclient==2.1.1
pymysql==1.0.2
DingtalkChatbot==1.3.0
loguru==0.6.0
django-apscheduler==0.6.2
snapshot-phantomjs==0.0.3
pyecharts==2.0.1
opencv-python==4.7.0.68
numpy
django-cors-headers
kafka-python
cachetools
ldap3